+++ rumqttd/src/link/meters.rs
@@ -1,6 +1,6 @@
-use crate::router::{Event, GetMeter, Meter};
+use crate::router::{Event, Meter};
 use crate::ConnectionId;
-use flume::{Receiver, RecvError, RecvTimeoutError, SendError, Sender, TrySendError};
+use flume::{Receiver, RecvError, RecvTimeoutError, SendError, Sender, TryRecvError, TrySendError};
 
 #[derive(Debug, thiserror::Error)]
 pub enum LinkError {
@@ -14,55 +14,38 @@ pub enum LinkError {
     RecvTimeout(#[from] RecvTimeoutError),
     #[error("Timeout = {0}")]
     Elapsed(#[from] tokio::time::error::Elapsed),
+    #[error("Channel try_recv error")]
+    TryRecv(#[from] TryRecvError),
 }
 
 pub struct MetersLink {
-    pub(crate) meter_id: ConnectionId,
-    router_tx: Sender<(ConnectionId, Event)>,
-    router_rx: Receiver<(ConnectionId, Vec<Meter>)>,
+    router_rx: Receiver<Vec<Meter>>,
 }
 
 impl MetersLink {
     pub fn new(router_tx: Sender<(ConnectionId, Event)>) -> Result<MetersLink, LinkError> {
-        let (tx, rx) = flume::bounded(5);
+        let (tx, rx) = flume::bounded(100);
+
         router_tx.send((0, Event::NewMeter(tx)))?;
-        let (meter_id, _meter) = rx.recv()?;
-
-        let link = MetersLink {
-            meter_id,
-            router_tx,
-            router_rx: rx,
-        };
-
+        let link = MetersLink { router_rx: rx };
         Ok(link)
     }
 
     pub async fn init(router_tx: Sender<(ConnectionId, Event)>) -> Result<MetersLink, LinkError> {
-        let (tx, rx) = flume::bounded(5);
-        router_tx.send((0, Event::NewMeter(tx)))?;
-        let (meter_id, _meter) = rx.recv_async().await?;
+        let (tx, rx) = flume::bounded(100);
 
-        let link = MetersLink {
-            meter_id,
-            router_tx,
-            router_rx: rx,
-        };
-
+        router_tx.send_async((0, Event::NewMeter(tx))).await?;
+        let link = MetersLink { router_rx: rx };
         Ok(link)
     }
 
-    pub fn get(&self, meter: GetMeter) -> Result<Vec<Meter>, LinkError> {
-        self.router_tx
-            .send((self.meter_id, Event::GetMeter(meter)))?;
-        let (_meter_id, meter) = self.router_rx.recv()?;
-        Ok(meter)
+    pub fn recv(&self) -> Result<Vec<Meter>, LinkError> {
+        let o = self.router_rx.try_recv()?;
+        Ok(o)
     }
 
-    pub async fn fetch(&self, meter: GetMeter) -> Result<Vec<Meter>, LinkError> {
-        self.router_tx
-            .send_async((self.meter_id, Event::GetMeter(meter)))
-            .await?;
-        let (_meter_id, meter) = self.router_rx.recv_async().await?;
-        Ok(meter)
+    pub async fn next(&self) -> Result<Vec<Meter>, LinkError> {
+        let o = self.router_rx.recv_async().await?;
+        Ok(o)
     }
 }